+++ CHANGES.md
@@ -20,6 +20,9 @@ Other Changes:
 - Added openmp dependence during compile time to allow for the use of `prange`.
 - Moved `cyrk_ode`'s Runge-Kutta constants to a separate module `CyRK.rk`. 
 
+Bug Fixes:
+- Fixed issue (for `nbrk_ode`) where incorrect step size could be used due to bad minimum step check (see [issue 20](https://github.com/jrenaud90/CyRK/issues/20)).
+
 ### v0.4.0
 
 New Features